employee-details rename context folder entries
