+++ src/context/context-compose.tsx
@@ -1,11 +1,11 @@
 // deps
 import React from 'react'
 // helpers
-import AuthContainer from '_/context/auth-context'
-import EmployeeContainer from '_/context/employees-context'
-import SnackbarContainer from '_/context/snackbar-context'
-import ThemeContainer from '_/context/theme-context'
-import EmployeeByIdContainer from '_/context/employee-by-id-context'
+import AuthContainer from '_/context/auth'
+import EmployeeContainer from '_/context/employees'
+import SnackbarContainer from '_/context/snackbar'
+import ThemeContainer from '_/context/theme'
+import EmployeeByIdContainer from '_/context/employee-by-id'
 
 const ContextCompose: React.FC = ({ children }) => (
   <>
