+++ src/context/index.ts
@@ -1,9 +1,9 @@
 import ContextCompose from '_/context/context-compose'
-import { AuthContext } from '_/context/auth-context'
-import { ThemeContext } from '_/context/theme-context'
-import { EmployeesContext } from '_/context/employees-context'
-import { EmployeeByIdContext } from '_/context/employee-by-id-context'
-import { SnackbarContext } from '_/context/snackbar-context'
+import { AuthContext } from '_/context/auth'
+import { ThemeContext } from '_/context/theme'
+import { EmployeesContext } from '_/context/employees'
+import { EmployeeByIdContext } from '_/context/employee-by-id'
+import { SnackbarContext } from '_/context/snackbar'
 
 export {
   ContextCompose as default,
